refactor: drop dead features/sandbox and sandbox-agent/ndjson-parser
Both modules were relics of the pre-daemon host-side architecture:
- `features/sandbox/materialization.ts` exported a dozen materialization
  helpers (materializeSummary, getDocsRoot, computeChecksum, etc.) that
  were superseded when the sandbox daemon took over materialization.
  Only `sanitizePathSegment` and `SyncLogger` were still consumed by
  sandbox-orchestration — both inlined where they're used.
- `features/sandbox-agent/ndjson-parser.ts` was for the host process
  parsing agent stdout. After PR #87 (sandbox daemon as sole agent
  path), sandbox-proxy.ts does its own NDJSON parsing inline; the
  exports were only consumed by their own test.

Four feature dirs collapse to three: chat, sandbox-agent (just prompt
builder now), sandbox-orchestration.
