+++ apps/chat-api/src/features/sandbox-agent/index.ts
@@ -1,4 +1,2 @@
-export type { AgentStreamEvent } from "./ndjson-parser";
-export { createNdjsonParser } from "./ndjson-parser";
 export type { SandboxAgentPromptOptions } from "./sandbox-agent.prompt";
 export { buildSandboxAgentPrompt } from "./sandbox-agent.prompt";
