--- apps/chat-api/src/features/sandbox/index.ts
@@ -1,13 +1,0 @@
-export {
-	type CollectionSymlink,
-	getCollectionDocsRoot,
-	getDocsRoot,
-	type MaterializationConfig,
-	type MaterializedFile,
-	materializeSummaries,
-	materializeSummary,
-	resolveCollectionSymlinks,
-	sanitizePathSegment,
-	type SummarySource,
-	type SyncLogger,
-} from "./materialization";
